% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_abide.R
\docType{data}
\name{ABIDE_0050002_funcSeg}
\alias{ABIDE_0050002_funcSeg}
\title{Subject 0050002 in ABIDE study, functional}
\format{A BrcFmri object of size 91 x 109 x 91 voxels, of which
there are 228,347 voxels in representing the brain (25.3%). The image is
not symmetric. The first 5 time slices are included.
\describe{
  \item{data2d}{5 x 228347 matrix, values representing the anatomical structure}
  \item{id}{The id, here "ABIDE_0050002"}
  \item{parcellation}{The singleton parcellation}
}}
\source{
Taken from the ABIDE study \url{http://fcon_1000.projects.nitrc.org/indi/abide/}
which is part of INDI \url{http://fcon_1000.projects.nitrc.org/}. The data was processed
using C-PAC \url{https://fcp-indi.github.io/}.
}
\description{
Subject 0050002 in ABIDE study, functional
}
\author{
Kevin Lin \email{kevinl1@andrew.cmu.edu}
}
\keyword{data}

