% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_abide.R
\docType{data}
\name{ABIDE_0050003_anat}
\alias{ABIDE_0050003_anat}
\title{Subject 0050003 in ABIDE study, anatomical}
\format{A BrcFmri object of size 91 x 109 x 91 voxels, of which
there are 262,876 voxels in representing the brain (29.1%). The image is
not symmetric.
\describe{
  \item{data2d}{1 x 262876 matrix, values representing the anatomical structure}
  \item{id}{The id, here "ABIDE_0050003"}
  \item{parcellation}{The singleton parcellation}
}}
\source{
Taken from the ABIDE study \url{http://fcon_1000.projects.nitrc.org/indi/abide/}
which is part of INDI \url{http://fcon_1000.projects.nitrc.org/}. The data was processed
using C-PAC \url{https://fcp-indi.github.io/}.
}
\description{
Subject 0050003 in ABIDE study, anatomical
}
\author{
Kevin Lin \email{kevinl1@andrew.cmu.edu}
}
\keyword{data}

