% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cobre.R
\docType{data}
\name{COBRE_0040071_funcSeg}
\alias{COBRE_0040071_funcSeg}
\title{Subject 0040071 in COBRE study, functional}
\format{A BrcFmri object of size 61 x 73 x 61 voxels, of which
there are 82,470 voxels in representing the brain (30.3%). The image is
not symmetric. The first 5 time slices are included.
\describe{
  \item{data2d}{1 x 82470 matrix, values representing the anatomical structure}
  \item{id}{The id, here "COBRE_0040071"}
  \item{parcellation}{The singleton parcellation}
}}
\source{
Taken from the COBRE study \url{http://fcon_1000.projects.nitrc.org/indi/abide/}
which is part of INDI \url{http://fcon_1000.projects.nitrc.org/}. The data was processed
using C-PAC \url{https://fcp-indi.github.io/}.
}
\description{
Subject 0040071 in COBRE study, functional
}
\author{
Kevin Lin \email{kevinl1@andrew.cmu.edu}
}
\keyword{data}

