% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_harvard.R
\docType{data}
\name{Harvard_Oxford}
\alias{Harvard_Oxford}
\title{Harvard Oxford Lateral Ventricles 2mm with Threshold 25}
\format{A BrcFmri object of size 91 x 109 x 91 voxels, of which
there are 2164 voxels in representing the brain (0.2%). The image is
not symmetric.
\describe{
  \item{data2d}{1 x 2164 matrix, values representing the anatomical structure}
  \item{id}{The id, here "Harvard_Oxford"}
  \item{parcellation}{The singleton parcellation}
}}
\source{
Taken from the FSL dataset.
}
\description{
Harvard Oxford Lateral Ventricles 2mm with Threshold 25
}
\author{
Kevin Lin \email{kevinl1@andrew.cmu.edu}
}
\keyword{data}

