% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MNI_2mm}
\alias{MNI_2mm}
\title{Montreal Neurological Institute brain with 2mm voxels}
\format{A BrcFmri object of size 91 x 109 x 91 voxels, of which
there are 228,453 voxels in representing the brain (25.3%).
\describe{
  \item{data2d}{1 x 228453 matrix, values representing the anatomical structure}
  \item{id}{The id, here "MNI_2mm"}
  \item{parcellation}{The singleton parcellation}
}}
\source{
Taken from the FSL dataset.
}
\description{
Montreal Neurological Institute brain with 2mm voxels
}
\author{
Kevin Lin \email{kevinl1@andrew.cmu.edu}
}
\keyword{data}

