% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mni.R
\docType{data}
\name{MNI_3mm}
\alias{MNI_3mm}
\title{Montreal Neurological Institute brain with 3mm voxels}
\format{A BrcFmri object of size 61 x 73 x 61 voxels, of which
there are 67,749 voxels in representing the brain (24.9%). The MNI image is
not symmetric.
\describe{
  \item{data2d}{1 x 67749 matrix, values representing the anatomical structure}
  \item{id}{The id, here "MNI_3mm"}
  \item{parcellation}{The singleton parcellation}
}}
\source{
Taken from the FSL dataset.
}
\description{
Montreal Neurological Institute brain with 3mm voxels
}
\author{
Kevin Lin \email{kevinl1@andrew.cmu.edu}
}
\keyword{data}

